pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "tola-data"
version = "1.0.50"
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
    # boto3 v1.36 produces an error on S3 storage which isn't Amazon AWS
    # If upgrading boto3 check that fix in `src/tola/s3client.py` is still
    # required.
    "boto3 ~= 1.36.0",
    "click",
    "duckdb",
    "mysql-connector-python",
    "psycopg2-binary", # PostgreSQL connector
    "pyarrow",
    "python-ulid",
    "pytz",
    "requests",
    'importlib-metadata; python_version >= "3.12"',
    "partisan",
    "tol-sdk>=1.8.8",
]

[project.scripts]
compare-pacbio-json="tola.compare_pacbio_json:cli"
data-processed="tola.data_processed:cli"
diff-mlwh="tola.diff_mlwh.diff_mlwh_cmd:cli"
duckdb-pacbio-report="tola.duckdb_pacbio_report:cli"
fetch-mlwh-seq-data="tola.fetch_mlwh_seq_data:cli"
fetch-ont-seq-data="tola.fetch_ont_seq_data:cli"
goat-client="tola.goat_client:cli"
jv="tola.tqc.view:view"
load-mlwh-seq-data="tola.load_mlwh_seq_data:main"
ncbi-taxonomy-duckdb="tola.ncbi_taxonomy_duckdb:cli"
pacbio-run-report="tola.pacbio_run_report:main"
status-duckdb="tola.status_sheets_to_duckdb:cli"
tqc="tola.tqc.tqc_cmd:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning:botocore"]

[tool.ruff.lint]
select = [
    "A",
    "ARG",
    "B",
    "BLE",
    "C4",
    "COM",
    "DTZ",
    "E",
    "EM",
    "F",
    "I",
    "ISC",
    "LOG",
    "N",
    "PYI",
    "S",
    "SIM",
    "U",
]
ignore = [
    "COM812",
    "C901",
]

[tool.ruff.lint.per-file-ignores]
"{test_*,*_test,conftest}.py" = ["E501", "S101"]

[tool.uv.sources]
partisan = { git = "https://github.com/wtsi-npg/partisan", rev = "3.3.0" }

[dependency-groups]
dev = [
    "SQLAlchemy ~= 2.0.0",
    "pytest",
    "ruff",
]
